// types/enums.ts
export enum SelectedOptions1 {
  WEAK = 'słabo',
  AVERAGE = 'przeciętnie',
  GOOD = 'dobrze',
  VERY_GOOD = 'bardzo dobrze',
}

export enum SelectedOptions2 {
  YES = 'tak',
  NO = 'nie',
  PROBABLY = 'raczej nie',
  DOESNT_MATTER = 'nie ma znaczenia',
}

export enum SelectedOptions3 {
  INSTRUCTIONS = 'instruktaże',
  TESTS = 'testy',
  ALL = 'wszystko',
  DOESNT_MATTER = 'nie ma znaczenia',
}

export enum SelectedOptions4 {
  COMMUNICATOR = 'komunikator',
  BLOG = 'blog medyczny',
  ASSISTANT = 'asystent AI',
  NEWS = 'newsy medyczne',
}

export enum SelectedOptions5 {
  FREE = 'za darmo',
  HUNDRED = '100',
  TWO_HUNDRED = '250',
  MONTHLY = '19.99',
}